Refactor: 타입에러 때문에 다시 페이지 각각에 작성해줌
