--- src/pages/project/components/ProjectForm/index.ts
@@ -1,2 +1,0 @@
-export { ProjectForm } from "./ProjectForm";
-export type { ProjectCreateFormType, ProjectEditFormType } from "./ProjectForm.types";
