+++ src/pages/project/edit/edit.types.ts
@@ -1,0 +1,13 @@
+export interface ProjectEditFormType {
+  name: string;
+  imageUrl: File;
+  projectType: string;
+  description: string;
+  positions: string[];
+  recruitCount: string;
+  languages: string[];
+  skills: string[];
+  areas: string[];
+  startDate: Date | null | "미정";
+  endDate: Date | null | "미정";
+}
